requirements.txt
Pillow>=9.0.0
flask>=2.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
